changed int's to size_t's

+++ ft_bzero.c
@@ -6,18 +6,18 @@
 /*   By: jpceia <jpceia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 21:42:48 by jpceia            #+#    #+#             */
-/*   Updated: 2021/02/08 21:42:49 by jpceia           ###   ########.fr       */
+/*   Updated: 2021/02/09 21:37:58 by jpceia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_bzero(void *str, int n)
+void	*ft_bzero(void *str, size_t n)
 {
 	char	*ptr;
 
 	ptr = (char*)str;
-	while (ptr - (char *)str < n)
+	while ((size_t)(ptr - (char *)str) < n)
 		*(ptr++) = 0;
 	return (str);
 }
