+++ ft_memcpy.c
@@ -6,15 +6,17 @@
 /*   By: jpceia <jpceia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 21:47:24 by jpceia            #+#    #+#             */
-/*   Updated: 2021/02/08 21:51:03 by jpceia           ###   ########.fr       */
+/*   Updated: 2021/02/09 21:38:18 by jpceia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	char	*ptr_dest;
 	char	*ptr_src;
-	int		index;
+	size_t	index;
 
 	ptr_dest = (char *)dest;
 	ptr_src = (char *)src;
