+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jpceia <jpceia@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/08 21:36:00 by jpceia            #+#    #+#             */
-/*   Updated: 2021/02/09 15:46:42 by jpceia           ###   ########.fr       */
+/*   Updated: 2021/02/09 21:47:45 by jpceia           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (str[index])
